requirements-web.txt
# Web Backend Dependencies for Scholia
# Phase A: FastAPI Backend

# Web Framework
fastapi
uvicorn[standard]

# AI/ML
anthropic

# PDF Processing
pymupdf

# File Upload
python-multipart

# Configuration
pydantic-settings
python-dotenv

# Database
aiosqlite

# Zotero Integration
pyzotero

# Notion Integration
notion-client
aiohttp
